Finish removal of SCN, RFE, and Discovery
As we convert SMD to a daemon that can run completely separately from the hardware it manages, we need to remove code that presumes locality and replace it with other functionality that does something similar.

For discovery, we're going with Magellan which is run manually.  We can evolve it to be a continuous operator in the future.

For State Change Notifications and Redfish Event Subscriptions, we have disabled both completely.  If we need to bring them back in the future, GitHub has a long memory.