Generalize container image build to target other CPU architectures
With this change, the target architecture for the container image can be
selected with environment variables in the build environment, rather than being
hard-coded into the Makefile. If the `GOOS` and `GOARCH` environment variables
are unspecified, the container image is built to target `linux/amd64`, which is
the same behavior as before.

This has been tested with `linux/arm64`, which takes around 15 minutes to build
the SMD image on a modern x86-based server.